backend/requirements.txt
django
psycopg
psycopg2-binary
djangorestframework-simplejwt
djangorestframework
